+++ site/kcworks/ext.py
@@ -28,13 +28,14 @@ from invenio_rdm_records.services.communities.components import (
     CommunityServiceComponents,
 )
 from invenio_rdm_records.services.components import DefaultRecordsComponents
+from invenio_rdm_records.services.config import FileServiceConfig
 from invenio_remote_user_data_kcworks.errors import (
-    BrokerTokenMissingError,
-    BrokerTokenDecryptionError,
-    BrokerPayloadExpiredError,
     BrokerExpiryValueError,
     BrokerNonceValidationError,
+    BrokerPayloadExpiredError,
     BrokerPayloadProcessingError,
+    BrokerTokenDecryptionError,
+    BrokerTokenMissingError,
     IDTokenInvalid,
     NoIDPFoundError,
     StateTokenInvalid,
@@ -59,6 +60,9 @@ from kcworks.services.records.community_inclusion.community_access import (
 )
 from kcworks.services.records.components.first_record_component import (
     FirstRecordComponent,
+)
+from kcworks.services.records.components.sanitize_filenames_component import (
+    FileNameSanitizerComponent,
 )
 from kcworks.services.records.components.per_field_permissions_component import (
     PerFieldEditPermissionsComponent,
@@ -170,6 +174,21 @@ class KCWorks:
             *existing_rdm_record_components,
             FirstRecordComponent,
             PerFieldEditPermissionsComponent,
+        ]
+
+        existing_record_file_components = app.config.get(
+            "RDM_FILES_SERVICE_COMPONENTS", FileServiceConfig.components
+        )
+        existing_draft_file_components = app.config.get(
+            "RDM_DRAFT_FILES_SERVICE_COMPONENTS", FileServiceConfig.components
+        )
+        app.config["RDM_FILES_SERVICE_COMPONENTS"] = [
+            FileNameSanitizerComponent,
+            *existing_record_file_components,
+        ]
+        app.config["RDM_DRAFT_FILES_SERVICE_COMPONENTS"] = [
+            FileNameSanitizerComponent,
+            *existing_draft_file_components,
         ]
 
         existing_record_communities_components = app.config.get(